requirements.txt
streamlit==1.33.0
pandas==2.2.2
numpy==1.26.4
faiss-cpu==1.7.4
sentence-transformers==2.6.1
scikit-learn==1.4.2
twilio==9.0.0
openai==1.30.0
tiktoken==0.7.0
python-dotenv==1.0.1
openpyxl==3.1.2